+++ CHANGELOG.md
@@ -1,0 +1,16 @@
+# Changelog
+
+All notable changes to the eTracer project will be documented in this file.
+
+## [0.1.0] - 2025-08-06
+
+### Added
+- Initial release of etracer package
+- Package structure with proper Python packaging standards
+- Enhanced stack trace formatting with colors
+- AI-powered error analysis using OpenAI API
+- Multiple usage modes (decorator, context manager, global handler)
+- Local variable inspection at error points
+- Smart caching system to reduce API calls
+- Customizable verbosity levels
+- Clean separation of interfaces and implementations
