+++ examples.py
@@ -3,8 +3,8 @@ Test script for the AI-enhanced tracer module.
 This demonstrates how to use the AI-powered error analysis feature.
 """
 
-import tracer
 import os
+import etracer
 
 # Configure AI integration (use your own API key)
 # You can also set this via environment variable: export OPENAI_API_KEY="your-api-key"
@@ -12,17 +12,17 @@ API_KEY = os.environ.get("OPENAI_API_KEY", "")
 
 if API_KEY:
     # Enable tracer with AI integration
-    tracer.configure_ai(api_key=API_KEY, model="gpt-3.5-turbo")
-    tracer.enable(verbosity=2, show_locals=True, ai_enabled=True)
+    etracer.configure_ai(api_key=API_KEY, model="gpt-3.5-turbo")
+    etracer.enable(verbosity=2, show_locals=True, ai_enabled=True)
     print("AI-powered error analysis enabled")
 else:
     # Enable tracer without AI
-    tracer.enable(verbosity=2, show_locals=True)
+    etracer.enable(verbosity=2, show_locals=True)
     print("AI analysis disabled (no API key provided)")
     print("Set the OPENAI_API_KEY environment variable to enable AI analysis")
 
 
-@tracer.debug
+@etracer.debug
 def test_division_by_zero():
     """Test zero division error with AI analysis"""
     print("Testing division by zero...")
@@ -32,7 +32,7 @@ def test_division_by_zero():
     return result
 
 
-@tracer.debug
+@etracer.debug
 def test_attribute_error():
     """Test attribute error with AI analysis"""
     print("Testing attribute error...")
@@ -41,7 +41,7 @@ def test_attribute_error():
     return x
 
 
-@tracer.debug
+@etracer.debug
 def test_index_error():
     """Test index error with AI analysis"""
     print("Testing index error...")
@@ -53,7 +53,7 @@ def test_index_error():
 def test_with_context_manager():
     """Test error handling with context manager"""
     print("Testing error with context manager...")
-    with tracer.catch_errors():
+    with etracer.catch_errors():
         # This will raise a TypeError
         result = "5" + 5
         print(f"Result: {result}")  # This line won't execute
