+++ src/etracer/utils/__init__.py
@@ -1,0 +1,22 @@
+"""
+Utility modules for etracer package.
+"""
+
+from .printer import Colors, ConsolePrinter, NoOpPrinter
+from .cache import CacheConfig, FileBasedCache
+from .timer import Timer
+from .spinner import Spinner, NoOpProgressIndicator
+from .ai_client import AIConfig, AIClient
+
+__all__ = [
+    "Colors",
+    "ConsolePrinter",
+    "NoOpPrinter",
+    "CacheConfig",
+    "FileBasedCache",
+    "Timer",
+    "Spinner",
+    "NoOpProgressIndicator",
+    "AIConfig",
+    "AIClient",
+]
