+++ tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Test suite for etracer package.
+"""
